+++ app/Http/Controllers/LapanganController.php
@@ -10,6 +10,10 @@ class LapanganController extends Controller
 {
     public function index()
     {
+        if(!auth()->user()->can('view lapangan')){
+            abort(403, 'Unauthorized action.');
+        }
+
         $lapangan = Lapangan::paginate(5);
 
         // return response()->json([
@@ -25,11 +29,19 @@ class LapanganController extends Controller
 
     public function create()
     {
-        return view('Lapangan.create');
+        if(!auth()->user()->can('create lapangan')){
+            abort(403, 'Unauthorized action.');
+        }
+        
+        return view('lapangan.create');
     }
 
     public function store(Request $request)
     {
+        if(!auth()->user()->can('create lapangan')){
+            abort(403, 'Unauthorized action.');
+        }
+        
         $validatedRequest =  $request->validate([
             'nama' => 'required|string|max:255',
             'tanggal' => 'date|required'
@@ -43,11 +55,15 @@ class LapanganController extends Controller
         //     'data' => $data
         // ], 201);
 
-        return redirect()->route('Lapangan.index')->with('success', 'Lapangan Berhasil ditambahkan');
+        return redirect()->route('lapangan.index')->with('success', 'Lapangan Berhasil ditambahkan');
     }
 
     public function show(Lapangan $lapangan)
     {
+        if(!auth()->user()->can('view jadwal')){
+            abort(403, 'Unauthorized action.');
+        }
+        
         $jadwal = $lapangan->jadwal()->with('user')->orderBy('jam_mulai')->paginate(5);
 
         // return response()->json([
@@ -59,7 +75,7 @@ class LapanganController extends Controller
         //     ]
         // ], 200);
 
-        return view('Lapangan.show', [
+        return view('lapangan.show', [
             'jadwal' => $jadwal,
             'lapangan' => $lapangan,
         ]);
@@ -67,19 +83,27 @@ class LapanganController extends Controller
 
     public function edit(Lapangan $lapangan)
     {
+        if(!auth()->user()->can('edit lapangan')){
+            abort(403, 'Unauthorized action.');
+        }
+        
         // return response()->json([
         //     'success'=> true,
         //     'message' => 'Data Lapangan berhasil dikirim ke halaman edit',
         //     'data' => $lapangan
         // ], 200);
 
-        return view('Lapangan.edit', [
+        return view('lapangan.edit', [
             'lapangan' => $lapangan
         ]);
     }
 
     public function update(Lapangan $lapangan, Request $request)
     {
+        if(!auth()->user()->can('edit lapangan')){
+            abort(403, 'Unauthorized action.');
+        }
+        
         $validatedRequest =  $request->validate([
             'nama' => 'required|string|max:255',
             'tanggal' => 'date|required'
@@ -93,11 +117,15 @@ class LapanganController extends Controller
         //     'data' => $lapangan 
         // ], 200);
 
-        return redirect()->route('Lapangan.index')->with('success', 'Lapangan Berhasil diubah');
+        return redirect()->route('lapangan.index')->with('success', 'Lapangan Berhasil diubah');
     }
 
     public function destroy(Lapangan $lapangan)
     {
+        if(!auth()->user()->can('delete lapangan')){
+            abort(403, 'Unauthorized action.');
+        }
+        
         $lapangan->delete();
 
         // return response()->json([
@@ -106,6 +134,6 @@ class LapanganController extends Controller
         //     'data' => $lapangan
         // ], 200);
 
-        return redirect()->route('Lapangan.index')->with('success', 'Berhasil menghapus lapangan');
+        return redirect()->route('lapangan.index')->with('success', 'Berhasil menghapus lapangan');
     }
 }
